rename 'dummy' test app too
